refactor(Parser): extract parsers to own files
